split exception class files
