+++ includes/WebServerException/ConfigException.hpp
@@ -1,0 +1,22 @@
+#pragma once
+
+#include <string>
+
+#include "WebServerException/WebServerException.hpp"
+
+class ConfigException : public WebServerException {
+ public:
+  class FileNotFound : public WebServerException {
+   private:
+    std::string message;
+
+   public:
+    FileNotFound(const char *);
+    virtual const char *what() const throw();
+  };
+  ConfigException();
+  virtual const char *what() const throw();
+
+ private:
+  std::string message;
+};
