+++ includes/WebServerException/ParsingException.hpp
@@ -1,0 +1,14 @@
+#pragma once
+
+#include <string>
+
+#include "WebServerException/WebServerException.hpp"
+
+class ParsingException : public WebServerException {
+ public:
+  ParsingException();
+  virtual const char *what() const throw();
+
+ private:
+  std::string message;
+};
