+++ includes/WebServerException/WebServerException.hpp
@@ -1,0 +1,17 @@
+#pragma once
+
+#include <exception>
+#include <string>
+
+class WebServerException : public std::exception {
+ public:
+  WebServerException();
+  virtual const char *what() const throw();
+
+ private:
+  std::string message;
+};
+
+#include "WebServerException/ConfigException.hpp"
+#include "WebServerException/HostingException.hpp"
+#include "WebServerException/ParsingException.hpp"
