+++ includes/webserver.hpp
@@ -4,4 +4,6 @@
 #include <iostream>
 #include <string>
 
+#include "WebServerException/WebServerException.hpp"
+
 #endif /* ***************************************************** WEBSERV_H */
